# src/api/swagger/portal-api.yaml
swagger: '2.0'
info:
  version: '1.0.0'
  title: wicked.haufe.io Portal API
  description: The API which the API Portal uses to store and retrieve information
 
# Under securityDefinitions we declare which security schemes can be used.
# These definitions do not affect to the operations, but can be applied
# using the 'security' key at the global level or the operation level.
# In this sample, we'll see how to apply the security definition globally,
# and override it at the operation level.
#
# Note that for Access Code Flow, you must provide the Token URL.
securityDefinitions:
  # For personalized access to data
  accessCode:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/auth/local/portal-api/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/auth/local/portal-api/oauth2/authorize
    flow: accessCode
    scopes:
      read_plans: Read plans
      read_groups: Read user group definitions
      read_content: Read custom/markdown/HTML content
      health: Read out health information
      read_templates: Read templates
      read_auth_servers: Read authorization servers
      read_apis: Read API definitions
      read_users: Read user information
      write_users: Write user information
      read_applications: Read application data
      write_applications: Write application data
      read_subscriptions: Read subscriptions
      write_subscriptions: Write/edit/delete subscriptions
      read_approvals: Read pending approvals
      read_auditlog: Read audit log
      read_verifications: Read email and password verifications
      write_verifications: Create or delete verifications
      read_registrations: Read user registrations
      write_registrations: Write user registrations
      read_grants: Read user API grants
      write_grants: Write user API grants
      webhooks: Access and change webhooks
      restart_api: Restart the API components
      access_tokens: Administer access tokens
  
  implicit:
    type: oauth2
    # This should be updated to the Token URL of your OAuth provider.
    tokenUrl: https://api.portal.com/portal-api/v1/oauth2/token
    # This should be updated to the Authorization URL of your OAuth provider.
    authorizationUrl: https://api.portal.com/portal-auth/api/portal-api
...
produces:
  - application/hal+json
host: 'localhost:3001'
basePath: /portal-api/v1
...
paths:
...
  '/approvals':
    get:
      security:
      - accessCode: ['read_approvals']
      - implicit: ['read_approvals']
      - password: ['read_approvals']
      tags:
        - misc
      description: Gets all pending approvals (admin)
      responses:
        '200':
          description: Successful request
          schema:
            type: array
            items:
              $ref: '#/definitions/approvalType'
  '/approvals/{approvalId}':
    get:
      security:
      - accessCode: ['read_approvals']
      - implicit: ['read_approvals']
      - password: ['read_approvals']
      tags:
        - misc
      description: Gets all pending approvals (admin)
      parameters:
        - name: approvalId
          in: path
          type: string
          required: true
          description: The ID of the approval record to retrieve
      responses:
        '200':
          description: Successful request
          schema:
            type: array
            items:
              $ref: '#/definitions/approvalType'
        '403':
          description: Not allowed
        '404':
          description: Not found
  '/approvals/byDateRange/{startDate}/{endDate}':
    get:
      security:
      - accessCode: ['read_approvals']
      - implicit: ['read_approvals']
      - password: ['read_approvals']
      tags:
        - misc
      description: Gets all pending approvals (admin)
      parameters:
        - name: startDate
          in: path
          type: string
          required: true
          description: The Start Date of the approval record to retrieve
          x-parameter-encoding: plain
        - name: endDate
          in: path
          type: string
          required: true
          description: The End Date of the approval record to retrieve
          x-parameter-encoding: plain
      responses:
        '200':
          description: Successful request
          schema:
            type: array
            items:
              $ref: '#/definitions/approvalType'
        '403':
          description: Not allowed
        '404':
          description: Not found
    delete:
      security:
      - accessCode: ['write_subscriptions']
      - implicit: ['write_subscriptions']
      - password: ['write_subscriptions']
      tags:
        - misc
      description: Delete all pending approvals (admin)
      parameters:
        - name: startDate
          in: path
          type: string
          required: true
          description: The Start Date of the approval records to delete
          x-parameter-encoding: plain
        - name: endDate
          in: path
          type: string
          required: true
          description: The End Date of the approval records to delete
          x-parameter-encoding: plain
      responses:
        '204':
          description: 'Successfully deleted, No content'
        '403':
          description: 'Forbiden, Not allowed'
        '404':
          description: 'Not Found'
...
definitions:
...
  approvalType:
    type: object
    properties:
      user:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
          email:
            type: string
      api:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
      application:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
      plan:
        type: object
        properties:
          id:
            type: string
          name:
            type: string
...